+++ compiler-tests/src/test/data/box/inject/InjectedFunctionParametersWithLambdaDefaultsWork.kt
@@ -1,0 +1,20 @@
+// Regression test to ensure we reparent copied lambda parameters
+class ExampleClass @Inject constructor(@Assisted private val lambdaParam: (String) -> Unit = {}) {
+  @AssistedFactory
+  interface Factory {
+    fun create(@Assisted lambdaParam: (String) -> Unit): ExampleClass
+  }
+}
+
+@DependencyGraph
+interface AppGraph {
+  val exampleClassFactory: ExampleClass.Factory
+}
+
+fun box(): String {
+  val graph = createGraph<AppGraph>()
+  val factory = graph.exampleClassFactory
+  val instance = factory.create {}
+  assertTrue(instance is ExampleClass)
+  return "OK"
+}